Revert "Make Prima a module to match CAF"
Sorry, I know this is annoying.  I thought it had to be a module to match the newer firmware, but apparently it doesn't.  Works fine thanks to @bbelos :thumbsup:

CAF updates prima fairly regularly, so in order to remain in sync with those updates, we are going to keep it built in.

This reverts commit db296387feabb82c35dc413c6a85924927ef45ba.